refactored links to be cleaner
